scripts/requirements.txt
# Phase 0 capture scripts dependencies
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
psycopg2-binary>=2.9.9
msal>=1.28.0
requests>=2.31.0
